requirements.txt
asgiref==3.7.2
Django==4.2.6
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.1
pillow==10.2.0
python-dotenv==1.0.1
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
